feat(fsd): A2 E1 — reconciliation cluster を entities へ (#376)
型のみ移設（Strategy R・挙動保存）。4 entity slice を新設し model.ts に
現行 interface を verbatim 移設・index.ts barrel を追加:
- entities/bank-transaction  (BankTransaction)
- entities/bank-import-batch (BankImportBatch)
- entities/reconciliation    (Reconciliation + ReconciliationAllocation)
- entities/client-credit     (ClientCredit)

consumer 再ポイント: endpoints.ts の型 block と 4 page（reconciliation は
2 entity へ分離）。ReconciliationAllocation は外部 importer 無し＝slice 内同梱。

logic-diff-0 証明: types/index.ts は 5 interface の削除のみ（+0 行）・移設 5
interface body は md5 一致（無改変）。model.ts の新規行は header doc comment
のみ（transitional 註記→設計 §4.1）。npm run check 緑（vitest 62 不変）。

Refs #376
