+++ frontend/src/entities/bank-import-batch/index.ts
@@ -1,0 +1,1 @@
+export type { BankImportBatch } from './model'
