+++ frontend/src/entities/bank-transaction/index.ts
@@ -1,0 +1,1 @@
+export type { BankTransaction } from './model'
