+++ frontend/src/entities/client-credit/index.ts
@@ -1,0 +1,1 @@
+export type { ClientCredit } from './model'
