+++ frontend/src/entities/reconciliation/index.ts
@@ -1,0 +1,1 @@
+export type { Reconciliation, ReconciliationAllocation } from './model'
