+++ frontend/src/pages/bank-import/BankImportPage.tsx
@@ -2,7 +2,7 @@ import { useState, useRef } from 'react'
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { listBankImportBatches, importBankCsv, reverseBankImportBatch, getClearSettings } from '@/shared/api/endpoints'
 import { describeApiError } from '@/shared/api/client'
-import type { BankImportBatch } from '@/types'
+import type { BankImportBatch } from '@/entities/bank-import-batch'
 import { Button } from '@/shared/ui/button'
 import { Card, CardHead, CardBody } from '@/shared/ui/card'
 import { DataTable, TableStateRow, SortableTh, nextSort, Pager } from '@/shared/ui/data-table'
