+++ frontend/src/pages/bank-transactions/BankTransactionsPage.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { useQuery } from '@tanstack/react-query'
 import { listBankTransactions, bankTransactionsExportPath } from '@/shared/api/endpoints'
 import { downloadCsv } from '@/shared/lib/download'
-import type { BankTransaction } from '@/types'
+import type { BankTransaction } from '@/entities/bank-transaction'
 import { Button } from '@/shared/ui/button'
 import { Card } from '@/shared/ui/card'
 import { DataTable, TableStateRow, Pager, SortableTh, nextSort } from '@/shared/ui/data-table'
