+++ frontend/src/pages/client-credits/ClientCreditsPage.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { listClientCredits, applyClientCredit, clientCreditsExportPath } from '@/shared/api/endpoints'
 import { downloadCsv } from '@/shared/lib/download'
-import type { ClientCredit } from '@/types'
+import type { ClientCredit } from '@/entities/client-credit'
 import { Button } from '@/shared/ui/button'
 import { Card } from '@/shared/ui/card'
 import { DataTable, TableStateRow, SortableTh, nextSort, Pager } from '@/shared/ui/data-table'
