+++ frontend/src/types/index.ts
@@ -10,70 +10,6 @@ export interface User {
   fiscal_year_end_month?: number | null
 }
 
-export interface BankImportBatch {
-  bank_import_batch_id: number
-  organization_id: number
-  bank_account_id: number
-  file_hash: string
-  source_filename: string
-  row_count: number
-  status: 'imported' | 'reversed'
-  imported_at: string
-  imported_by: number
-  reversed_at: string | null
-  reversal_reason: string | null
-}
-
-export interface BankTransaction {
-  bank_transaction_id: number
-  organization_id: number
-  bank_import_batch_id: number
-  bank_account_id: number
-  value_date: string
-  amount_cents: number
-  counterparty_text: string
-  status: 'unmatched' | 'partially_matched' | 'matched' | 'voided'
-}
-
-export interface ReconciliationAllocation {
-  reconciliation_allocation_id: number
-  organization_id: number
-  payment_reconciliation_id: number
-  invoice_id: number
-  amount_cents: number
-  payment_id: number | null
-  external_reference: string | null
-}
-
-export interface Reconciliation {
-  payment_reconciliation_id: number
-  organization_id: number
-  bank_transaction_id: number
-  status: 'confirmed' | 'reversed'
-  reason_code: string | null
-  confirmed_by: number
-  confirmed_at: string
-  reversed_at: string | null
-  reversal_reason: string | null
-  allocations: ReconciliationAllocation[]
-}
-
-export interface ClientCredit {
-  client_credit_id: number
-  organization_id: number
-  client_id: number | null
-  client_name: string | null
-  source: 'invoice_upstream' | 'manual'
-  manual_receivable_id: number | null
-  amount_cents: number
-  remaining_cents: number
-  status: 'open' | 'voided'
-  source_bank_transaction_id: number
-  reconciliation_id: number
-  created_by: number
-  created_at: string
-}
-
 /**
  * A receivable entered directly in Clear, not sourced from NeNe Invoice
  * (ADR 0014). A reconciliation reference — NOT an issued invoice / 適格請求書 /
